requirements.txt
streamlit
plotly
pandas
duckdb
pyarrow
google-genai
requests
